Centralize MLDB connection management + generalize random name creation
Instead of hardcoding the mldb connection in the code, there is now
a new method set_connection in the connection module that must be called
first. The rest of the module will use that information.

In a few places, especially when predicting something, a random name was
generated. Now the user can specify the name or if None,
generate_random_name is called